after refactor

+++ hotel-entities/src/main/java/by/nc/teamone/entities/models/RoomModel.java
@@ -9,14 +9,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class RoomModel{
     public int coast;
     public int seats;
-    public int status_id;
-    public int type_id;
+    public long status_id;
+    public long type_id;
 
     @JsonCreator
     public RoomModel(@JsonProperty("coast")int coast,
                      @JsonProperty("seats")int seats,
-                     @JsonProperty("status_id")int status_id,
-                     @JsonProperty("type_id")int type_id) {
+                     @JsonProperty("status_id")long status_id,
+                     @JsonProperty("type_id")long type_id) {
 
         this.coast = coast;
         this.seats = seats;
